[Fix] RemoveListener, property
이벤트 처리에서 이벤트 다가갈 때 set프로퍼티에 연결해야하는데 실수,
RemoveListener 미적용,
값전달시 변수 실수

+++ Assets/Scripts/UI/UIGuage.cs
@@ -18,7 +18,7 @@ public class UIGuage : MonoBehaviour
     {
         Debug.Log($"UI : {value}");
         bagGuageImage.fillAmount = value;
-        int bagValue = (int)value * Manager.InvenInstance.maxSum;
+        float bagValue = value * Manager.InvenInstance.maxSum;
         bagText.text = bagValue.ToString();
     }
